refactor(postgresql): clean up
