+++ app/requirements.txt
@@ -2,20 +2,16 @@ streamlit>=1.41.1,<2.0.0
 psycopg2-binary>=2.9.9,<3.0.0
 python-dotenv>=1.0.1,<2.0.0
 pandas>=2.2.3,<3.0.0
-pgvector>=0.2.5,<0.3.0
 requests>=2.31.0,<3.0.0
 pydantic>=2.0.0,<3.0.0
 anthropic>=0.8.1
 pytest>=7.4.3,<8.0.0
-sqlalchemy>=2.0.0,<3.0.0
 numpy>=1.26.0,<2.0.0
 beautifulsoup4>=4.12.2,<5.0.0
-thefuzz
 nest-asyncio
 torch
 crawl4ai>=0.5.0,<0.6.0rc1
 pydantic>=2.0.0,<3.0.0
-ollama
 openai
 spacy
 markdown
@@ -27,7 +23,6 @@ langsmith>=0.1.0,<0.2.0
 langchain-core>=0.1.0,<0.2.0
 streamlit-star-ratings
 streamlit-tags
-plotly
 openpyxl
 xlsxwriter
 pyexcel_ods3
