+++ app/utils/crawler_utils.py
@@ -1,6 +1,6 @@
 import logging
 import sys
-from typing import List, Optional
+from typing import List
 
 import nest_asyncio
