+++ app/🚎_Tags extraction - output.py
